fix array getSettingValue bug

+++ R/getSettingValue.R
@@ -30,7 +30,7 @@ getSettingValue <- function(key,settings){
     if(typeof(value)=="list"){
       value<-getSettingValue(key[2:length(key)],value)  
     #If position is provided and the value is a vector
-    }else if(typeof(key[[2]])=="double" & length(value)>=key[[2]] & length(key)==2){ 
+    }else if(typeof(key[[2]])=="integer" & length(value)>=key[[2]] & length(key)==2){ 
       value<-value[[key[[2]]]]
     }else{
       #If there are more keys, but the value is not a list, return NULL
